backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.352
langchain-community==0.0.6
torch==2.1.1
transformers==4.35.2
python-docx==1.0.1
google-cloud-translate==3.12.1
sentence-transformers==2.2.2
pydantic==2.5.2
faiss-cpu==1.7.4
pypdf==3.17.1 